Support python 2.7 only on legacy
3.4 is in separate branch, 3.5+ is mainstream

* no cythonize code
* no "magic" branches of code
* typehints backported
