+++ .travis.yml
@@ -2,13 +2,8 @@ sudo: false
 language: python
 os: linux
 python:
-- 2.7
-- 3.4
-- 3.5
-- 3.6
-- &mainstream_python 3.7-dev
+- &mainstream_python 2.7=
 - &pypy pypy
-- pypy3.5
 install:
 - &upgrade_python_toolset pip install --upgrade pip setuptools wheel
 - pip install tox-travis
@@ -48,8 +43,7 @@ jobs:
     - docker
     install:
     - *upgrade_python_toolset
-    script:
-    - ./tools/run_docker.sh "logwrap"
+    script: []
     before_deploy:
     - pip install -r build_requirements.txt
     - python setup.py bdist_wheel
