+++ requirements.txt
@@ -1,2 +1,4 @@
 six >=1.9
 typing >= 3.6 ; python_version < "3.8"
+funcsigs>=1.0 ; python_version == "2.7"
+enum34>=1.1 ; python_version == "2.7"
