+++ setup.py
@@ -20,21 +20,9 @@ from __future__ import print_function
 
 import ast
 import collections
-from distutils.command import build_ext
-import distutils.errors
-import glob
 import os.path
-import shutil
-import sys
-
-try:
-    from Cython.Build import cythonize
-except ImportError:
-    cythonize = None
 
 import setuptools
-
-PY3 = sys.version_info[:2] > (2, 7)
 
 with open(
     os.path.join(
@@ -49,84 +37,6 @@ with open('requirements.txt') as f:
 
 with open('README.rst',) as f:
     long_description = f.read()
-
-
-def _extension(modpath):
-    """Make setuptools.Extension."""
-    return setuptools.Extension(modpath, [modpath.replace('.', '/') + '.py'])
-
-
-requires_optimization = [
-    _extension('logwrap._class_decorator'),
-    _extension('logwrap._log_wrap_shared'),
-    _extension('logwrap._repr_utils'),
-]
-
-if 'win32' != sys.platform:
-    requires_optimization.append(
-        _extension('logwrap.__init__')
-    )
-
-ext_modules = cythonize(
-    requires_optimization,
-    compiler_directives=dict(
-        always_allow_keywords=True,
-        binding=True,
-        embedsignature=True,
-        overflowcheck=True,
-        language_level=3,
-    )
-) if cythonize is not None and PY3 else []
-
-
-class BuildFailed(Exception):
-    """For install clear scripts."""
-    pass
-
-
-class AllowFailRepair(build_ext.build_ext):
-    """This class allows C extension building to fail and repairs init."""
-
-    def run(self):
-        """Run."""
-        try:
-            build_ext.build_ext.run(self)
-
-            # Copy __init__.py back to repair package.
-            build_dir = os.path.abspath(self.build_lib)
-            root_dir = os.path.abspath(os.path.join(__file__, '..'))
-            target_dir = build_dir if not self.inplace else root_dir
-
-            src_files = (
-                os.path.join('logwrap', '__init__.py'),
-                # _log_wrap3 should not be compiled due to specific bug:
-                # Exception inside `async def` crashes python.
-                os.path.join('logwrap', '_log_wrap3.py'),
-            )
-
-            for src_file in src_files:
-                src = os.path.join(root_dir, src_file)
-                dst = os.path.join(target_dir, src_file)
-
-                if src != dst:
-                    shutil.copyfile(src, dst)
-        except (
-            distutils.errors.DistutilsPlatformError,
-            getattr(globals()['__builtins__'], 'FileNotFoundError', OSError)
-        ):
-            raise BuildFailed()
-
-    def build_extension(self, ext):
-        """build_extension."""
-        try:
-            build_ext.build_ext.build_extension(self, ext)
-        except (
-            distutils.errors.CCompilerError,
-            distutils.errors.DistutilsExecError,
-            distutils.errors.DistutilsPlatformError,
-            ValueError
-        ):
-            raise BuildFailed()
 
 
 # noinspection PyUnresolvedReferences
@@ -180,8 +90,6 @@ def get_simple_vars_from_src(src):
         ast.Str, ast.Num,
         ast.List, ast.Set, ast.Dict, ast.Tuple
     )
-    if PY3:
-        ast_data += (ast.Bytes, ast.NameConstant,)
 
     tree = ast.parse(src)
 
@@ -225,11 +133,6 @@ classifiers = [
 
     'Programming Language :: Python :: 2',
     'Programming Language :: Python :: 2.7',
-    'Programming Language :: Python :: 3',
-    'Programming Language :: Python :: 3.4',
-    'Programming Language :: Python :: 3.5',
-    'Programming Language :: Python :: 3.6',
-    'Programming Language :: Python :: 3.7',
 
     'Programming Language :: Python :: Implementation :: CPython',
     'Programming Language :: Python :: Implementation :: PyPy',
@@ -256,7 +159,7 @@ setup_args = dict(
     long_description=long_description,
     classifiers=classifiers,
     keywords=keywords,
-    python_requires='>=2.7.5,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*',
+    python_requires='>=2.7.5,<3.0',
     # While setuptools cannot deal with pre-installed incompatible versions,
     # setting a lower bound is not harmful - it makes error messages cleaner. DO
     # NOT set an upper bound on setuptools, as that will lead to uninstallable
@@ -266,35 +169,10 @@ setup_args = dict(
     setup_requires="setuptools >= 21.0.0,!=24.0.0,"
                    "!=34.0.0,!=34.0.1,!=34.0.2,!=34.0.3,!=34.1.0,!=34.1.1,!=34.2.0,!=34.3.0,!=34.3.1,!=34.3.2,"
                    "!=36.2.0",
-    extras_require={
-        ':python_version == "2.7"': [
-            'funcsigs>=1.0',
-            'enum34>=1.1',
-        ],
-    },
     install_requires=required,
     package_data={
-        'logwrap': [
-            os.path.basename(filename)
-            for filename in glob.glob(os.path.join('logwrap', '*.pyi'))
-        ] + [
-            'py.typed'
-        ],
+        'logwrap': ['py.typed'],
     },
 )
-if PY3 and cythonize is not None:
-    setup_args['ext_modules'] = ext_modules
-    setup_args['cmdclass'] = dict(build_ext=AllowFailRepair)
 
-try:
-    setuptools.setup(**setup_args)
-except BuildFailed:
-    print(
-        '*' * 80 + '\n'
-        '* Build Failed!\n'
-        '* Use clear scripts version.\n'
-        '*' * 80 + '\n'
-    )
-    del setup_args['ext_modules']
-    del setup_args['cmdclass']
-    setuptools.setup(**setup_args)
+setuptools.setup(**setup_args)
