+++ test/test_log_wrap_shared.py
@@ -19,20 +19,13 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-import sys
 import unittest
 
-import six
-
-# noinspection PyProtectedMember
-from logwrap import _log_wrap_shared
+import logwrap
 
 # pylint: disable=ungrouped-imports, no-name-in-module
-if six.PY3:  # pragma: no cover
-    from inspect import signature
-else:  # pragma: no cover
-    # noinspection PyUnresolvedReferences
-    from funcsigs import signature
+# noinspection PyUnresolvedReferences
+from funcsigs import signature
 # pylint: enable=ungrouped-imports, no-name-in-module
 
 
@@ -48,7 +41,7 @@ sig = signature(example_function)
 # noinspection PyUnusedLocal,PyMissingOrEmptyDocstring
 class TestBind(unittest.TestCase):
     def test_001_positive(self):
-        params = list(_log_wrap_shared.bind_args_kwargs(sig, 1, arg3=33))
+        params = list(logwrap.bind_args_kwargs(sig, 1, arg3=33))
         arg_1_bound = params[0]
         self.assertEqual(arg_1_bound.name, 'arg1')
         self.assertEqual(arg_1_bound.value, 1)
@@ -90,7 +83,7 @@ class TestBind(unittest.TestCase):
         self.assertEqual(str(kwargs_bound), "**kwargs={}")
 
     def test_002_args_kwargs(self):
-        params = list(_log_wrap_shared.bind_args_kwargs(sig, 1, 2, 3, 4, arg5=5))
+        params = list(logwrap.bind_args_kwargs(sig, 1, 2, 3, 4, arg5=5))
 
         args_bound = params[3]
         self.assertEqual(args_bound.name, 'args')
@@ -109,40 +102,8 @@ class TestBind(unittest.TestCase):
         self.assertEqual(str(kwargs_bound), "**kwargs={'arg5': 5}")
 
     def test_003_no_value(self):
-        params = list(_log_wrap_shared.bind_args_kwargs(sig, 1, arg3=33))
+        params = list(logwrap.bind_args_kwargs(sig, 1, arg3=33))
         arg_1_bound = params[0]
         arg1_parameter = arg_1_bound.parameter
         with self.assertRaises(ValueError):
-            _log_wrap_shared.BoundParameter(arg1_parameter, arg1_parameter.empty)
-
-    @unittest.skipIf(sys.version_info[:2] < (3, 4), 'python 3 syntax')
-    def test_004_annotations(self):
-        namespace = {}
-        exec("""def func(arg1, arg2: int, arg3: int=3): pass""", namespace)
-        func = namespace['func']
-        sig = signature(func)
-        params = list(_log_wrap_shared.bind_args_kwargs(sig, 1, 2, 4))
-
-        arg_1_bound = params[0]
-        self.assertEqual(arg_1_bound.name, 'arg1')
-        self.assertEqual(arg_1_bound.value, 1)
-        self.assertEqual(arg_1_bound.default, arg_1_bound.empty)
-        self.assertEqual(arg_1_bound.annotation, arg_1_bound.empty)
-        self.assertEqual(arg_1_bound.kind, arg_1_bound.POSITIONAL_OR_KEYWORD)
-        self.assertEqual(str(arg_1_bound), "arg1=1")
-
-        arg_2_bound = params[1]
-        self.assertEqual(arg_2_bound.name, 'arg2')
-        self.assertEqual(arg_2_bound.value, 2)
-        self.assertEqual(arg_2_bound.default, arg_2_bound.empty)
-        self.assertEqual(arg_2_bound.annotation, int)
-        self.assertEqual(arg_2_bound.kind, arg_2_bound.POSITIONAL_OR_KEYWORD)
-        self.assertEqual(str(arg_2_bound), "arg2: int=2")
-
-        arg_3_bound = params[2]
-        self.assertEqual(arg_3_bound.name, 'arg3')
-        self.assertEqual(arg_3_bound.value, 4)
-        self.assertEqual(arg_3_bound.default, 3)
-        self.assertEqual(arg_3_bound.annotation, int)
-        self.assertEqual(arg_3_bound.kind, arg_3_bound.POSITIONAL_OR_KEYWORD)
-        self.assertEqual(str(arg_3_bound), "arg3: int=4  # 3")
+            logwrap.BoundParameter(arg1_parameter, arg1_parameter.empty)
