+++ test/test_repr_utils.py
@@ -21,7 +21,6 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-import sys
 import unittest
 
 import logwrap
@@ -286,115 +285,3 @@ class TestPrettyRepr(unittest.TestCase):
 
     def test_py2_compatibility_flag(self):
         self.assertIsInstance(logwrap.pretty_repr(u'Text', py2_str=True), str)
-
-
-# noinspection PyUnusedLocal,PyMissingOrEmptyDocstring
-@unittest.skipIf(
-    sys.version_info[:2] < (3, 4),
-    'Strict python 3.3+ API'
-)
-class TestAnnotated(unittest.TestCase):
-    def test_001_annotation_args(self):
-        fmt = "\n{spc:<{indent}}<{obj!r} with interface ({args}){annotation}>".format
-        namespace = {}
-
-        exec("""
-import typing
-def func(a: typing.Optional[int]=None):
-    pass
-                        """,
-             namespace
-             )
-        func = namespace['func']  # type: typing.Callable[..., None]
-
-        self.assertEqual(
-            logwrap.pretty_repr(func),
-            fmt(
-                spc='',
-                indent=0,
-                obj=func,
-                args="\n    a: typing.Union[int, NoneType]=None,\n",
-                annotation=""
-            )
-        )
-
-        self.assertEqual(
-            logwrap.pretty_str(func),
-            fmt(
-                spc='',
-                indent=0,
-                obj=func,
-                args="\n    a: typing.Union[int, NoneType]=None,\n",
-                annotation=""
-            )
-        )
-
-    def test_002_annotation_return(self):
-        fmt = "\n{spc:<{indent}}<{obj!r} with interface ({args}){annotation}>".format
-        namespace = {}
-
-        exec("""
-import typing
-def func() -> None:
-    pass
-                                """,
-             namespace
-             )
-        func = namespace['func']  # type: typing.Callable[[], None]
-
-        self.assertEqual(
-            logwrap.pretty_repr(func),
-            fmt(
-                spc='',
-                indent=0,
-                obj=func,
-                args='',
-                annotation=' -> None'
-            )
-        )
-
-        self.assertEqual(
-            logwrap.pretty_str(func),
-            fmt(
-                spc='',
-                indent=0,
-                obj=func,
-                args='',
-                annotation=' -> None'
-            )
-        )
-
-    def test_003_complex(self):
-        fmt = "\n{spc:<{indent}}<{obj!r} with interface ({args}){annotation}>".format
-        namespace = {}
-
-        exec("""
-import typing
-def func(a: typing.Optional[int]=None) -> None:
-    pass
-                                """,
-             namespace
-             )
-        func = namespace['func']  # type: typing.Callable[..., None]
-
-        self.assertEqual(
-            logwrap.pretty_repr(func),
-            fmt(
-                spc='',
-                indent=0,
-                obj=func,
-                args="\n    a: typing.Union[int, NoneType]=None,\n",
-                annotation=" -> None"
-            )
-        )
-
-        self.assertEqual(
-            logwrap.pretty_str(func),
-            fmt(
-                spc='',
-                indent=0,
-                obj=func,
-                args="\n    a: typing.Union[int, NoneType]=None,\n",
-                annotation=" -> None"
-            )
-        )
