+++ tox.ini
@@ -5,7 +5,7 @@
 
 [tox]
 minversion = 2.0
-envlist = pep8, pep257, py{27,34,35,36,37,py,py3}, pylint, docs, bandit, py{34,35,36}-nocov, mypy
+envlist = pep8, pep257, py{27,py}, pylint, docs, bandit, mypy
 skipsdist = True
 skip_missing_interpreters = True
 
@@ -28,38 +28,12 @@ commands =
     py.test -vv --junitxml=unit_result.xml --html=report.html --cov-config .coveragerc --cov-report html --cov=logwrap {posargs:test}
     coverage report --fail-under 87
 
-[testenv:py34-nocov]
-usedevelop = False
-commands =
-    python setup.py bdist_wheel
-    pip install logwrap --no-index -f dist
-    py.test -vv {posargs:test}
-
-[testenv:py35-nocov]
-usedevelop = False
-commands =
-    python setup.py bdist_wheel
-    pip install logwrap --no-index -f dist
-    py.test -vv {posargs:test}
-
-[testenv:py36-nocov]
-usedevelop = False
-commands =
-    python setup.py bdist_wheel
-    pip install logwrap --no-index -f dist
-    py.test -vv {posargs:test}
-
 [testenv:venv]
 commands = {posargs:}
 
 [tox:travis]
 2.7 = install, py27,
-3.4 = py34,
-3.5 = py35,
-3.6 = py36,
-3.7 = install, py37,
 pypy = install, pypy,
-pypy3 = install, pypy3,
 
 [testenv:pep8]
 deps =
@@ -79,12 +53,14 @@ usedevelop = False
 commands = pip install ./ -vvv -U
 
 [testenv:pylint]
+usedevelop = False
 deps =
     pylint<2
     -r{toxinidir}/CI_REQUIREMENTS.txt
 commands = pylint logwrap
 
 [flake8]
+usedevelop = False
 exclude =
     .venv,
     .git,
@@ -103,11 +79,13 @@ count = True
 max-line-length = 120
 
 [testenv:docs]
+usedevelop = False
 deps =
     sphinx
 commands = python setup.py build_sphinx
 
 [testenv:bandit]
+usedevelop = False
 deps = bandit
 commands = bandit -r logwrap
 
@@ -120,6 +98,7 @@ commands =
     pipenv graph
 
 [testenv:mypy]
+usedevelop = False
 deps =
     mypy>=0.620
     -r{toxinidir}/CI_REQUIREMENTS.txt
